Move stuff back to the root folder
