--- Web-Project-Spring-Semester-2025/README.md
@@ -1,1 +1,0 @@
-# Web-Project-Spring-Semester-2025
