--- Web-Project-Spring-Semester-2025/browse.html
@@ -1,88 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="dark">
-    <head>
-        <meta charset="UTF-8" />
-        <link rel="icon" type="image/x-icon" href="favicon.ico">
-        <title>NextFit</title>
-        <link href="/style.css" rel="stylesheet">
-        <script type="module" src="/main.js"></script>
-        <script type="module" src="/src/browse.js"></script>
-    </head>
-    <body>
-        <div class ="flex flex-col gap-2 layout">
-            <header class="navbar p-4 backdrop-blur-md">
-                <div class="flex flex-wrap gap-4 items-center justify-between">
-                  <a href="/">
-                    <img src="assets/Logo.png" width="128" />
-                  </a>
-                  <div>
-                    <div class="justify-start hidden lg:flex gap-3" id="popUpMenu">
-                      <a href="./browse.html" class="nav-link">Browse Exercises</a>
-                    </div>
-                  </div>
-                  <form action="browse.html" class="ml-auto max-lg:hidden flex gap-1" id="searchForm">
-                    <input type="search" class="input" placeholder="Search" name="query" />
-                    <button class="btn btn-small hover:cursor-pointer" onclick="document.querySelector('searchForm').submit()">
-                      <img src="assets/MdiMagnify.svg" width="24"/>
-                    </button>
-                  </form>
-                  <div class="gap-3 flex max-lg:hidden" id="loginRegister">
-                    <a id="login" href="login.html" class="nav-link">Login</a>
-                    <a id="register" href="register.html" class="nav-link">Register</a>
-                    <a id="btnTheme" href="#" class="nav-link">🌙</a>
-                  </div>
-                  <div class="gap-3 flex max-lg:hidden" id="userInfo" style="display: none">
-                    <img id="userPicture" src="assets/default-avatar.png" width="40" height="40" />
-                    <button id="logoutButton" class="nav-link">Logout</button>
-                  </div>
-                  <button class="lg:hidden nav-link" id="btn-menu">
-                    <img src="assets/MdiMenu.svg" />
-                  </button>
-                </div>
-                <div class="relative w-full flex flex-col items-end">
-                  <div id="mobile-menu" class="hidden lg:hidden flex flex-col gap-3 px-4 py-3 content rounded shadow-md w-full">
-                    <a href="./browse.html" class="nav-link">Browse Exercises</a>
-                    <a href="login.html" class="nav-link">Login</a>
-                    <a href="register.html" class="nav-link">Register</a>
-                  </div>
-                </div>
-            </header>
-            <main>
-                <!--Edit here mostly, the rest should remain the same-->
-                <strong class="text-3xl">Browse Exercises</strong>
-                <div class="flex grow gap-4">
-                    <div class="content grow self-start flex flex-col gap-4" style="flex: 1;">
-                        <strong class="text-3xl mx-auto">Filters</strong>
-                        <div class="flex gap-2">
-                            <div class="flex flex-col gap-1 grow">
-                                <label for="exercises-query">Search</label>
-                                <input class="input" id="exercises-query">
-                            </div>
-                            <div class="flex flex-col gap-1 grow">
-                                <label for="exercises-diff">Difficulty</label>
-                                <select class="input" id="exercises-diff">
-                                    <option value="any">Any</option>
-                                    <option value="beginner">Beginner</option>
-                                    <option value="intermediate">Intermediate</option>
-                                    <option value="advanced">Advanced</option>
-                                </select>
-                            </div>
-                        </div>
-                        <div class="body flex gap-4 mx-auto" style="min-height: 450px;">
-                            <img src="/assets/body.png" usemap="#body-map" onmouseover="highlightPart()" onclick="alert('Muscle group mannequin functionality to be implemented.');">
-                            <map name="body-map">
-                                <area shape="rect" coords="49,104,76,149" onmouseover="highlightPart('biceps')"/>
-                                <area shape="rect" coords="174,104,198,148" onmouseover="highlightPart('biceps')"/>
-                            </map>
-                            <div class="body-parts">
-                                <img src="/assets/Biceps.png" id="biceps" class="opacity-0" style="width: 250px; top: 15px; left: 1px;">
-                            </div>
-                        </div>
-                    </div>
-                    <div class="flex flex-col grow gap-4" style="flex: 2;" id="exercise-list"></div>
-                </div>
-            </main>
-            <footer class="content">Made By Group 2 - Braude College</footer>
-        </div>
-    </body>
-</html>
