--- Web-Project-Spring-Semester-2025/data/users.json
@@ -1,16 +1,0 @@
-[
-    {
-        "username":"Moran",
-        "email":"morkupfer@braude.ac.il",
-        "password":"lala1234",
-        "dob": "10/01/2000",
-        "picture": "assets/default-avatar.png"
-    },
-    {
-        "username":"Ilya",
-        "email":"IlyaZ@braude.ac.il",
-        "password":"bubu1234",
-        "dob": "20/05/2000",
-        "picture": "assets/default-avatar.png"
-    }
-]
