--- Web-Project-Spring-Semester-2025/exercise.html
@@ -1,73 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="dark">
-    <head>
-        <meta charset="UTF-8" />
-        <link rel="icon" type="image/x-icon" href="favicon.ico">
-        <title>NextFit</title>
-        <link href="/style.css" rel="stylesheet">
-        <script type="module" src="/main.js"></script>
-        <script type="module" src="src/exercise.js"></script>
-    </head>
-    <body>
-            <div class ="flex flex-col gap-2 layout">
-            <header class="navbar p-4 backdrop-blur-md">
-                <div class="flex flex-wrap gap-4 items-center justify-between">
-                <a href="/">
-                    <img src="assets/Logo.png" width="128" />
-                </a>
-                <div>
-                    <div class="justify-start hidden lg:flex gap-3" id="popUpMenu">
-                    <a href="./browse.html" class="nav-link">Browse Exercises</a>
-                    </div>
-                </div>
-                <form action="browse.html" class="ml-auto max-lg:hidden flex gap-1" id="searchForm">
-                    <input type="search" class="input" placeholder="Search" name="query" />
-                    <button class="btn btn-small hover:cursor-pointer" onclick="document.querySelector('searchForm').submit()">
-                    <img src="assets/MdiMagnify.svg" width="24"/>
-                    </button>
-                </form>
-                <div class="gap-3 flex max-lg:hidden" id="loginRegister">
-                    <a id="login" href="login.html" class="nav-link">Login</a>
-                    <a id="register" href="register.html" class="nav-link">Register</a>
-                    <a id="btnTheme" href="#" class="nav-link">🌙</a>
-                </div>
-                <div class="gap-3 flex max-lg:hidden" id="userInfo" style="display: none">
-                    <img id="userPicture" src="assets/default-avatar.png" width="40" height="40" />
-                    <button id="logoutButton" class="nav-link">Logout</button>
-                </div>
-                <button class="lg:hidden nav-link" id="btn-menu">
-                    <img src="assets/MdiMenu.svg" />
-                </button>
-                </div>
-                <div class="relative w-full flex flex-col items-end">
-                <div id="mobile-menu" class="hidden lg:hidden flex flex-col gap-3 px-4 py-3 content rounded shadow-md w-full">
-                    <a href="./browse.html" class="nav-link">Browse Exercises</a>
-                    <a href="login.html" class="nav-link">Login</a>
-                    <a href="register.html" class="nav-link">Register</a>
-                </div>
-                </div>
-            </header>
-            <main class="flex-row!">
-                <!--Edit here mostly, the rest should remain the same-->
-                <iframe 
-                    style="aspect-ratio: 16 / 9; width: 70%;"
-                    id="exercise-video"
-                    src="https://www.youtube.com/embed/caXXntNqTAg"
-                    title="How to squat ✅"
-                    frameborder="0"
-                    allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-                    referrerpolicy="strict-origin-when-cross-origin" allowfullscreen>
-                </iframe>
-                <div class="content p-4 w-2/12 overflow-hidden" style= "flex: 1;">
-                    <strong class="text-2xl " id="exercise-title"></strong>
-                    <br>
-                    <span id="exercise-difficulty"></span>
-                    <br>
-                    <br>
-                    <p class="whitespace-break-spaces" id="exercise-description"></p>
-                </div>
-            </main>
-            <footer class="p-4 content rounded">Made By Group 2 - Braude College</footer>
-        </div>
-    </body>
-</html>
