--- Web-Project-Spring-Semester-2025/index.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="dark">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/x-icon" href="favicon.ico" />
-    <title>NextFit</title>
-    <link href="/style.css" rel="stylesheet" />
-    <script type="module" src="/main.js"></script>
-  </head>
-  <body>
-    <div class="relative min-h-screen flex flex-col">
-      <div class="absolute inset-0 -z-10">
-        <img src="assets/hero.jpg" alt="Hero Background" class="w-full h-full object-cover" />
-        <div class="absolute inset-0 bg-black/60"></div>
-      </div>
-
-      <header class="navbar p-4 backdrop-blur-md">
-        <div class="flex flex-wrap gap-4 items-center justify-between">
-          <a href="/">
-            <img src="assets/Logo.png" width="128" />
-          </a>
-          <div>
-            <div class="justify-start hidden lg:flex gap-3" id="popUpMenu">
-              <a href="./browse.html" class="nav-link">Browse Exercises</a>
-            </div>
-          </div>
-          <form action="browse.html" class="ml-auto max-lg:hidden flex gap-1" id="searchForm">
-            <input type="search" class="input" placeholder="Search" name="query" />
-            <button class="btn btn-small hover:cursor-pointer" onclick="document.querySelector('searchForm').submit()">
-              <img src="assets/MdiMagnify.svg" width="24"/>
-            </button>
-          </form>
-          <div class="gap-3 flex max-lg:hidden" id="loginRegister">
-            <a id="login" href="login.html" class="nav-link">Login</a>
-            <a id="register" href="register.html" class="nav-link">Register</a>
-            <a id="btnTheme" href="#" class="nav-link">🌙</a>
-          </div>
-          <div class="gap-3 flex max-lg:hidden" id="userInfo" style="display: none">
-            <img id="userPicture" src="assets/default-avatar.png" width="40" height="40" />
-            <button id="logoutButton" class="nav-link">Logout</button>
-          </div>
-          <button class="lg:hidden nav-link" id="btn-menu">
-            <img src="assets/MdiMenu.svg" />
-          </button>
-        </div>
-        <div class="relative w-full flex flex-col items-end">
-          <div id="mobile-menu" class="hidden lg:hidden flex flex-col gap-3 px-4 py-3 content rounded shadow-md w-full">
-            <a href="./browse.html" class="nav-link">Browse Exercises</a>
-            <a href="login.html" class="nav-link">Login</a>
-            <a href="register.html" class="nav-link">Register</a>
-          </div>
-        </div>
-      </header>
-
-      <main class="grow flex flex-col items-center w-full">
-        <section class="flex flex-col items-center justify-center text-center text-white px-4 py-32 z-10">
-          <h1 class="text-4xl md:text-5xl font-bold mb-4">Welcome to NextFit</h1>
-          <p class="text-lg md:text-xl mb-6">Train smart. Stay fit. Anywhere.</p>
-          <a href="browse.html" class="btn primary px-6 py-3 rounded text-lg transition">
-            Browse Exercises
-          </a>
-        </section>
-
-        <div class="px-4 my-8 flex flex-col items-center text-white">
-          <strong class="text-3xl">What is NextFit?</strong>
-          <div class="text-xl text-center max-w-xl mt-2">
-            A site for watching exercise clips so you can get the best out of you!
-          </div>
-        </div>
-      </main>
-
-      <footer class="p-4 content rounded z-10">
-        Made By Group 2 - Braude College
-      </footer>
-    </div>
-  </body>
-</html>
