--- Web-Project-Spring-Semester-2025/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "myfirstapp",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tailwindcss/forms": "^0.5.10",
-    "@tailwindcss/typography": "^0.5.16",
-    "autoprefixer": "^10.4.20",
-    "prettier": "3.5.3",
-    "tailwindcss": "^4.1.1",
-    "vite": "^6.2.0"
-  },
-  "dependencies": {
-    "@tailwindcss/vite": "^4.1.1"
-  }
-}
