--- Web-Project-Spring-Semester-2025/vite.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite'
-import tailwindcss from '@tailwindcss/vite'
-export default defineConfig({
-  plugins: [
-    tailwindcss(),
-  ],
-})
