+++ index.html
@@ -7,14 +7,14 @@
     <link href="/style.css" rel="stylesheet" />
     <script type="module" src="/main.js"></script>
   </head>
-  <body class="bg-[var(--bg-color)] text-[var(--text-color)]">
+  <body>
     <div class="relative min-h-screen flex flex-col">
       <div class="absolute inset-0 -z-10">
         <img src="assets/hero.jpg" alt="Hero Background" class="w-full h-full object-cover" />
         <div class="absolute inset-0 bg-black/60"></div>
       </div>
 
-      <header class="navbar p-4 fixed w-full top-0 z-50 bg-black/60 backdrop-blur-md">
+      <header class="navbar p-4 backdrop-blur-md">
         <div class="flex flex-wrap gap-4 items-center justify-between">
           <a href="/">
             <img src="assets/Logo.png" width="128" />
@@ -26,8 +26,8 @@
           </div>
           <form action="browse.html" class="ml-auto max-lg:hidden flex gap-1" id="searchForm">
             <input type="search" class="input" placeholder="Search" name="query" />
-            <button class="btn-small hover:cursor-pointer" onclick="document.querySelector('searchForm').submit()">
-              <img src="assets/MdiMagnify.svg" width="24" />
+            <button class="btn btn-small hover:cursor-pointer" onclick="document.querySelector('searchForm').submit()">
+              <img src="assets/MdiMagnify.svg" width="24"/>
             </button>
           </form>
           <div class="gap-3 flex max-lg:hidden" id="loginRegister">
@@ -39,24 +39,24 @@
             <img id="userPicture" src="assets/default-avatar.png" width="40" height="40" />
             <button id="logoutButton" class="nav-link">Logout</button>
           </div>
-        </div>
-        <div class="relative w-full flex flex-col items-end">
-          <button class="lg:hidden mb-2" id="btnMenu">
+          <button class="lg:hidden nav-link" id="btn-menu">
             <img src="assets/MdiMenu.svg" />
           </button>
-          <div id="mobileMenu" class="hidden lg:hidden flex flex-col gap-3 px-4 py-3 bg-[var(--content-bg-color)] rounded shadow-md w-48 text-right">
-            <a href="./browse.html" class="nav-link block px-4 py-2 rounded hover:bg-gray-700 transition">Browse Exercises</a>
-            <a href="login.html" class="nav-link block px-4 py-2 rounded hover:bg-gray-700 transition">Login</a>
-            <a href="register.html" class="nav-link block px-4 py-2 rounded hover:bg-gray-700 transition">Register</a>
+        </div>
+        <div class="relative w-full flex flex-col items-end">
+          <div id="mobile-menu" class="hidden lg:hidden flex flex-col gap-3 px-4 py-3 content rounded shadow-md w-full">
+            <a href="./browse.html" class="nav-link">Browse Exercises</a>
+            <a href="login.html" class="nav-link">Login</a>
+            <a href="register.html" class="nav-link">Register</a>
           </div>
         </div>
       </header>
 
-      <main class="grow flex flex-col items-center w-full pt-[88px]">
+      <main class="grow flex flex-col items-center w-full">
         <section class="flex flex-col items-center justify-center text-center text-white px-4 py-32 z-10">
           <h1 class="text-4xl md:text-5xl font-bold mb-4">Welcome to NextFit</h1>
           <p class="text-lg md:text-xl mb-6">Train smart. Stay fit. Anywhere.</p>
-          <a href="browse.html" class="bg-red-600 hover:bg-red-700 text-white px-6 py-3 rounded text-lg transition">
+          <a href="browse.html" class="btn primary px-6 py-3 rounded text-lg transition">
             Browse Exercises
           </a>
         </section>
@@ -69,7 +69,7 @@
         </div>
       </main>
 
-      <footer class="p-4 content rounded z-10 text-white">
+      <footer class="p-4 content rounded z-10">
         Made By Group 2 - Braude College
       </footer>
     </div>
