+++ style.css
@@ -19,8 +19,12 @@ main {
     --text-color: #000;
     --bg-color: #fff;
     --content-bg-color: #e6e6e6;
-    --btn-bg-color: hsl(0, 0%, 30%);
     --input-bg-color: rgb(206, 206, 206);
+    --btn-bg-color: hsl(0, 0%, 30%);
+    --btn-hover-color: hsl(0, 0%, 35%);
+    --nav-hover-color: hsl(0, 0%, 80%);
+    --btn-primary-color: #f00303;
+    --btn-primary-hover-color: hsl(0, 98%, 40%);
     color-scheme: light;
 }
 
@@ -29,6 +33,8 @@ html.dark {
     --bg-color: hsl(0, 0%, 5%);
     --content-bg-color: hsl(0, 0%, 10%);
     --input-bg-color: #303030;
+    --nav-hover-color: hsl(0, 0%, 35%);
+
     color-scheme: dark;
 }
 
@@ -39,10 +45,28 @@ html.dark {
 }
 
 .btn {
+    color: white;
+    transition: background-color 0.25s;
     background-color: var(--btn-bg-color);
     padding: 0.5rem 1rem;
     border-radius: 0.5rem;
     text-align: center;
+}
+
+.btn.primary {
+    background-color: var(--btn-primary-color);
+}
+
+.btn.primary:hover {
+    background-color: var(--btn-primary-hover-color);
+}
+
+.btn img {
+    color: var(--text-color);
+}
+
+.btn:hover {
+    background-color: var(--btn-hover-color);
 }
 
 a:link {
@@ -78,6 +102,7 @@ select.input {
     background-color: var(--content-bg-color);
     position: sticky;
     line-height: 1;
+    z-index: 100;
     top: 0;
     left: 0;
 }
@@ -98,7 +123,7 @@ select.input {
     transition: opacity 0.25s;
 }
 
-#logoutButton {
+button {
     cursor: pointer;
 }
 
@@ -106,4 +131,18 @@ select.input {
     display: flex;
     gap: 12px;
     text-decoration: none !important;
+}
+
+.nav-link {
+    padding: 0.75rem;
+    transition: background-color 0.25s;
+    border-radius: 0.25rem;
+}
+
+.nav-link:hover {
+    background-color: var(--nav-hover-color);
+}
+
+#mobile-menu .nav-link{
+    width: 100%;
 }
